requirements.txt
numpy==1.22.2
Pillow==9.0.1